requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.1
grpcio==1.56.2
grpcio-tools==1.56.2
h11==0.14.0
idna==3.4
protobuf==4.23.4
pydantic==2.1.1
pydantic-core==2.4.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
